+++ renderer/rend/mesh.cpp
@@ -6,7 +6,7 @@ namespace rend
 {
 
 Mesh::Mesh(const vector<math::vec3> &vertices,
-           const vector<int> &indices,
+           const vector<size_t> &indices,
            const MeshType type)
     : m_vertices(vertices),
       m_indices(indices),
@@ -28,7 +28,7 @@ const vector<math::vec3> &Mesh::vertices() const
     return m_vertices;
 }
 
-const vector<int> &Mesh::indices() const
+const vector<size_t> &Mesh::indices() const
 {
     return m_indices;
 }
